+++ src/pandas_contract/_decorator.py
@@ -10,6 +10,7 @@ from pandas_contract.mode import Modes, get_mode
 from ._checks import (
     Check,
     CheckExtends,
+    CheckInplace,
     CheckKeepIndex,
     CheckKeepLength,
     CheckSchema,
@@ -49,6 +50,8 @@ class argument:  # noqa: N801
     :param same_size_as: The name of the input argument(s) that should have the same
         size.
     :param key: The key of the input to check. If None, the entire input is checked.
+    :param extends: Name of argument that this output extends.
+
 
     ## Examples:
     ### Ensure that input dataframe as a column "a" of type int.
@@ -186,6 +189,9 @@ class result:  # noqa: N801
         This can be used if the function returns a tuple or a mapping.
     :param extends: The name of the input argument that the output extends. Only
         columns defined in the output schema are allowed to be mutated or added.
+    :param inplace: Name of argument tha the output is identical to. The resulting
+        data-frame/series is identical to the given argument (assert fn(df) is df),
+        i.e. the function changes the dataframe in-place.
 
     ## Examples
     ### Ensure that the output dataframe has a column "a" of type int.
@@ -259,6 +265,7 @@ class result:  # noqa: N801
     same_size_as: str | Sequence[str] = ()
     key: Any = _UNDEFINED
     extends: str | None = None
+    inplace: str | None = None
 
     def __post_init__(self) -> None:
         self.same_index_as = same_index_as = ensure_list(self.same_index_as)
@@ -274,6 +281,7 @@ class result:  # noqa: N801
             CheckKeepIndex(same_index_as),
             CheckKeepLength(same_size_as),
             CheckExtends(self.extends, self.schema, "return"),
+            CheckInplace(self.inplace),
         ]
         self.checks: list[Check] = [check for check in checks if check.is_active]